feat(articles): Update ArticlesManager interface and implementation
