+++ app/src/androidTest/java/com/makentoshe/habrachan/AppActivityTest.kt
@@ -3,7 +3,7 @@ package com.makentoshe.habrachan
 import androidx.lifecycle.Lifecycle
 import androidx.test.core.app.ActivityScenario
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.makentoshe.habrachan.common.network.manager.HabrArticleManager
+import com.makentoshe.habrachan.common.network.manager.ArticlesManager
 import com.makentoshe.habrachan.di.common.ApplicationScope
 import com.makentoshe.habrachan.view.main.MainFlowFragment
 import okhttp3.OkHttpClient
@@ -25,8 +25,8 @@ class AppActivityTest : AndroidBaseTest() {
             val json = getRawJsonResource(R.raw.get_articles_success)
             val interceptor = ResponseInterceptor(200, json)
             val client = OkHttpClient.Builder().addInterceptor(interceptor).build()
-            val manager = HabrArticleManager.Builder(client).build()
-            bind<HabrArticleManager>().toInstance(manager)
+            val manager = ArticlesManager.Builder(client).build()
+            bind<ArticlesManager>().toInstance(manager)
         })
 
         val scenario = ActivityScenario.launch(AppActivity::class.java)
