+++ app/src/main/java/com/makentoshe/habrachan/viewmodel/article/VoteArticleViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.makentoshe.habrachan.common.database.SessionDao
 import com.makentoshe.habrachan.common.entity.article.VoteArticleResponse
-import com.makentoshe.habrachan.common.network.manager.HabrArticleManager
+import com.makentoshe.habrachan.common.network.manager.ArticlesManager
 import com.makentoshe.habrachan.common.network.request.VoteArticleRequest
 import io.reactivex.Observable
 import io.reactivex.android.schedulers.AndroidSchedulers
@@ -13,7 +13,7 @@ import io.reactivex.subjects.PublishSubject
 
 class VoteArticleViewModel(
     private val sessionDao: SessionDao,
-    private val articleManager: HabrArticleManager
+    private val articlesManager: ArticlesManager
 ) : ViewModel() {
 
     private val disposables = CompositeDisposable()
@@ -29,7 +29,7 @@ class VoteArticleViewModel(
     fun voteUp(articleId: Int) {
         val session = sessionDao.get()!!
         val request = VoteArticleRequest(session.clientKey, session.tokenKey, articleId)
-        articleManager.voteUp(request).subscribe(
+        articlesManager.voteUp(request).subscribe(
             voteArticleSubject::onNext, voteArticleErrorSubject::onNext
         ).let(disposables::add)
     }
@@ -37,7 +37,7 @@ class VoteArticleViewModel(
     fun voteDown(articleId: Int) {
         val session = sessionDao.get()!!
         val request = VoteArticleRequest(session.clientKey, session.tokenKey, articleId)
-        articleManager.voteDown(request).subscribe(
+        articlesManager.voteDown(request).subscribe(
             voteArticleSubject::onNext, voteArticleErrorSubject::onNext
         ).let(disposables::add)
     }
@@ -46,10 +46,10 @@ class VoteArticleViewModel(
 
     class Factory(
         private val sessionDao: SessionDao,
-        private val articleManager: HabrArticleManager
+        private val articlesManager: ArticlesManager
     ) : ViewModelProvider.NewInstanceFactory() {
         override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-            return VoteArticleViewModel(sessionDao, articleManager) as T
+            return VoteArticleViewModel(sessionDao, articlesManager) as T
         }
     }
 }
